requirements.txt
anyio @ file:///Users/runner/miniforge3/conda-bld/anyio_1621480459900/work/dist
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094662606/work
argon2-cffi @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi_1605217032873/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
Babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1619719576210/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1613413910603/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454858/work
click==8.0.1
conda==4.10.1
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1618231726924/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494586/work
cycler==0.10.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask==2.0.1
Flask-Cors==3.0.10
future @ file:///Users/runner/miniforge3/conda-bld/future_1610147429103/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1622488289347/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1620913002553/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1622679308382/work
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146802570/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1612125274649/work
jupyter-server @ file:///home/conda/feedstock_root/build_artifacts/jupyter_server_1621649383518/work
jupyterlab @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_1621368934875/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1622556871679/work
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1610099836705/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1621455750847/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1620506033308/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1610112888823/work
nbclassic @ file:///home/conda/feedstock_root/build_artifacts/nbclassic_1621947328608/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1614336084111/work
nbconvert @ file:///Users/runner/miniforge3/conda-bld/nbconvert_1605401843962/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1621259862661/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1622014657525/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.2.4
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
patsy==0.5.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1621288044168/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1616432837031/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1610094841871/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1620245170812/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1610146832352/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291460269/work
python-dateutil==2.8.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1622038487673/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1611943386799/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1619628632127/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1619561960622/work
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1611834504644/work
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio @ file:///Users/runner/miniforge3/conda-bld/sniffio_1610318406004/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1612273721995/work
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1621535250115/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
torch @ file:///Users/runner/miniforge3/conda-bld/pytorch-recipe_1619039080682/work
torchvision==0.10.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1610094701146/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1621890532941/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1615828766818/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client @ file:///Users/runner/miniforge3/conda-bld/websocket-client_1610127751547/work
Werkzeug==2.0.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
